requirements.txt
Django==5.1.2
djangorestframework==3.15.1
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.6.1
django-environ==0.11.2
django-cors-headers==4.4.0
django-redis==5.4.0
psycopg2-binary==2.9.9
redis==5.0.7
celery==5.4.0
python-dotenv==1.0.1
gunicorn==22.0.0
whitenoise==6.6.0
Pillow==10.4.0
python-slugify==8.0.4
